moved to other files and defined functions
